refactor: remove barrel files except for modules
