--- src/chains/evm/common/constants/abi/index.ts
@@ -1,1 +1,0 @@
-export * from "./erc-20-abi.js";
