--- src/chains/evm/common/constants/index.ts
@@ -1,1 +1,0 @@
-export * from "./chain.js";
