--- src/chains/evm/common/utils/index.ts
@@ -1,3 +1,0 @@
-export * from "./chain.js";
-export * from "./contract.js";
-export * from "./provider.js";
