--- src/chains/evm/hub/constants/abi/index.ts
@@ -1,2 +1,0 @@
-export * from "./account-manager-abi.js";
-export * from "./bridge-router-hub-abi.js";
