--- src/chains/evm/hub/constants/index.ts
@@ -1,1 +1,0 @@
-export * from "./chain.js";
