+++ src/chains/evm/hub/modules/folks-hub-loan.ts
@@ -1,27 +1,29 @@
+import { UINT256_LENGTH } from "../../../../common/constants/bytes.js";
+import { FINALITY } from "../../../../common/constants/message.js";
+import { Action } from "../../../../common/types/message.js";
+import { getRandomGenericAddress } from "../../../../common/utils/address.js";
+import { convertNumberToBytes } from "../../../../common/utils/bytes.js";
 import {
-  FINALITY,
-  UINT256_LENGTH,
-} from "../../../../common/constants/index.js";
-import { Action } from "../../../../common/types/index.js";
+  getSpokeChain,
+  getSpokeTokenData,
+} from "../../../../common/utils/chain.js";
 import {
   DEFAULT_MESSAGE_PARAMS,
   buildMessagePayload,
-  convertNumberToBytes,
-  getRandomGenericAddress,
   getSendTokenExtraArgsWhenRemoving,
-  getSpokeChain,
-  getSpokeTokenData,
-} from "../../../../common/utils/index.js";
+} from "../../../../common/utils/messages.js";
 import { getHubChain, getHubTokenData } from "../utils/chain.js";
 import { getBridgeRouterHubContract } from "../utils/contract.js";
 
 import type {
+  NetworkType,
   FolksChainId,
+} from "../../../../common/types/chain.js";
+import type {
   MessageAdapters,
   MessageToSend,
-  FolksTokenId,
-  NetworkType,
-} from "../../../../common/types/index.js";
+} from "../../../../common/types/message.js";
+import type { FolksTokenId } from "../../../../common/types/token.js";
 import type { Hex, PublicClient } from "viem";
 
 export function getSendTokenAdapterFees(
