--- src/chains/evm/hub/types/index.ts
@@ -1,3 +1,0 @@
-export * from "./account.js";
-export * from "./chain.js";
-export * from "./token.js";
