+++ src/chains/evm/hub/types/token.ts
@@ -1,8 +1,6 @@
-import type {
-  LoanType,
-  GenericAddress,
-  ITokenData,
-} from "../../../../common/types/index.js";
+import type { GenericAddress } from "../../../../common/types/chain.js";
+import type { LoanType } from "../../../../common/types/module.js";
+import type { ITokenData } from "../../../../common/types/token.js";
 
 export type HubTokenData = {
   poolId: number;
