--- src/chains/evm/hub/utils/index.ts
@@ -1,2 +1,0 @@
-export * from "./chain.js";
-export * from "./contract.js";
