--- src/chains/evm/spoke/constants/abi/index.ts
@@ -1,3 +1,0 @@
-export * from "./bridge-router-spoke-abi.js";
-export * from "./spoke-common-abi.js";
-export * from "./spoke-token-abi.js";
