+++ src/common/constants/chain.ts
@@ -1,9 +1,10 @@
 import { avalancheFuji, baseSepolia, sepolia } from "viem/chains";
 
-import { AdapterType, ChainType, NetworkType } from "../types/index.js";
+import { NetworkType, ChainType } from "../types/chain.js";
+import { AdapterType } from "../types/message.js";
 import { convertToGenericAddress } from "../utils/address.js";
 
-import type { FolksChain, FolksChainId, SpokeChain } from "../types/index.js";
+import type { FolksChainId, FolksChain, SpokeChain } from "../types/chain.js";
 
 export const MAINNET_FOLKS_CHAIN_ID = {} as const;
