--- src/common/constants/index.ts
@@ -1,3 +1,0 @@
-export * from "./bytes.js";
-export * from "./chain.js";
-export * from "./message.js";
