--- src/common/types/index.ts
@@ -1,5 +1,0 @@
-export * from "./chain.js";
-export * from "./core.js";
-export * from "./message.js";
-export * from "./module.js";
-export * from "./token.js";
