--- src/common/utils/index.ts
@@ -1,5 +1,0 @@
-export * from "./adapter.js";
-export * from "./address.js";
-export * from "./chain.js";
-export * from "./bytes.js";
-export * from "./messages.js";
