+++ src/index.ts
@@ -1,4 +1,17 @@
+// CORE
 export { FolksCore } from "./xchain/core/folks-core.js";
+
+// MODULES
 export { FolksAccount, FolksLoan } from "./xchain/modules/index.js";
-export * from "./common/types/index.js";
-export * from "./common/constants/index.js";
+
+// COMMON
+export * from "./common/types/address.js";
+export * from "./common/types/chain.js";
+export * from "./common/types/core.js";
+export * from "./common/types/message.js";
+export * from "./common/types/module.js";
+export * from "./common/types/token.js";
+
+export * from "./common/constants/bytes.js";
+export * from "./common/constants/chain.js";
+export * from "./common/constants/message.js";
